pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "mcp-servers"
version = "0.1.0"
description = "A project to learn MCP concepts, FastMCP, and FastAPI-MCP."
authors = [
    {name = "Vishal150494",email = "vchegde.hegde@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastmcp (>=2.13.1,<3.0.0)",
    "fastapi (>=0.123.0,<0.124.0)",
    "fastapi-mcp (>=0.4.0,<0.5.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "feedparser (>=6.0.12,<7.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
